backend/requirements.txt
flask==3.0.1
flask-cors==4.0.0
python-dotenv==1.0.0
anthropic==0.8.0
PyJWT==2.8.0
requests==2.31.0
pytest==7.4.4
black==24.1.0
flake8==7.0.0
gunicorn==21.2.0